Conflicts:
	src/javafxTest/Controller.java
	src/javafxTest/Main.java
	src/javafxTest/defaultShapes.java
	src/javafxTest/fxStation.java

+++ src/javafxTest/fxEndLine.java
@@ -27,7 +27,7 @@ public class fxEndLine extends Polyline {
         setStrokeLineCap(StrokeLineCap.ROUND);
         setStrokeLineJoin(StrokeLineJoin.ROUND);
 
-        int x = modelSt.getPosition().getX(), y=modelSt.getPosition().getY();
+        double x = modelSt.getPosition().getX(), y=modelSt.getPosition().getY();
 
         if(middleY == y) {
             if(middleX > x) {
